Remove CollapsedTrackingProtectionBadge
